pyproject.toml
[tool.poetry]
name = "home-monitoring"
version = "0.1.0"
description = ""
authors = ["alex-cryptobro <alex@cryptobro.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
python-telegram-bot = "^20.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
